pyproject.toml
[tool.poetry]
name = "boomer-graphics"
version = "0.1.0"
description = ""
authors = ["Shun-Fu Hu <husohome98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.6"
cloudinary = "^1.41.0"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"
openai = "^1.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
